Fix: harden story e.3 voice webhook test suite
